Add files via upload
This is pretty flushed out start. I included the SPACE(); map and revised the dialogue. I also fixed the WRDSRCH(); engine to work from vectors that are initialized above that function. The place for you to insert the password dialogue/puzzle is commented out. I will branch this branch when I finish the dialogue to assess damages before the function call to fix the engine!