Add files via upload
Here's the engine I made for the battle. I've been fine tuning it all day, so I don't think the player will likely win the first try, but I need to test it out on people and see what happens! Don't forget to change the address of "colors.h" when trying it out in puTTy.